content/media/webspeech/recognition/test/test_call_start_from_end_handler.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=650295
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 650295 -- Restart recognition from end handler</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=650295">Mozilla Bug 650295</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  function createAudioStream() {
    var audioTag = document.createElement("audio");
    audioTag.src = DEFAULT_AUDIO_SAMPLE_FILE;

    var stream = audioTag.mozCaptureStreamUntilEnded();
    audioTag.play();

    return stream;
  }

  function endHandler(evt, sr) {
    try {
      sr.start(createAudioStream()); // shouldn't fail
    } catch (err) {
      ok(false, "Failed to start() from end() callback");
    }

    info("Successfully start() from end() callback");
  }

  function expectExceptionHandler(evt, sr) {
    try {
      sr.start(createAudioStream());
    } catch (err) {
      is(err.name, "InvalidStateError");
      return;
    }

    ok(false, "Calling start() didn't raise InvalidStateError");
  }

  performTest({
    eventsToRequest: [
      'EVENT_RECOGNITIONSERVICE_FINAL_RESULT'
    ],
    expectedEvents: {
      'start': expectExceptionHandler,
      'audiostart': expectExceptionHandler,
      'speechstart': expectExceptionHandler,
      'speechend': expectExceptionHandler,
      'audioend': expectExceptionHandler,
      'result': buildResultCallback("Mock final result"),
      'end': endHandler,
    },
    doneFunc: SimpleTest.finish,
    prefs: [["media.webspeech.test.fake_fsm_events", true], ["media.webspeech.test.fake_recognition_service", true]]
  });

</script>
</pre>
</body>
</html>
